(Without Debugging) Add the main solution of Level-based gear compaction picking step.
We modified the listener.h to add the kGearCompactionAllInOne as a new compaction reason.

We modified the version_set.cc and version_set.h to record the biggest tree in VersionSetStorageInfo

We modified the Compaction_job.cc's install function, after the calling of InstallCompactionResults, and before LOGGING part, update the biggest tree information in the VersionStorageInfo.